feat: refactor user and userProfile validation schemas for improved organization and clarity
