+++ backend/src/validations/userProfile.validation.ts
@@ -1,18 +1,88 @@
 import { z } from "@/swagger/registry";
+import { user, userProfile } from "@/db/schema";
+import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import {
+  Certification,
   insertCertificationsSchema,
-  insertEducationsSchema,
-  insertUserProfileSchema,
-  insertWorkExperiencesSchema,
   selectCertificationsSchema,
-  selectEducationsSchema,
-  selectUserProfileSchema,
-  selectWorkExperiencesSchema,
   updateCertificationsSchema,
+} from "@/validations/certifications.validation";
+import {
+  Education,
+  insertEducationsSchema,
+  selectEducationsSchema,
   updateEducationsSchema,
-  updateUserProfileSchema,
+} from "@/validations/educations.validation";
+import {
+  insertWorkExperiencesSchema,
+  selectWorkExperiencesSchema,
   updateWorkExperiencesSchema,
-} from "@/db/schema";
+  WorkExperience,
+} from "@/validations/workExperiences.validation";
+
+// Zod schemas
+export const insertUserSchema = createInsertSchema(user, {
+  email: z.email("Invalid email format").toLowerCase(),
+  fullName: z.string().min(1, "First name is required").max(100).trim(),
+  status: z.enum(["active", "deactivated", "deleted"]).default("active"),
+  deletedAt: z.date().optional(),
+});
+
+export const insertUserProfileSchema = createInsertSchema(userProfile, {
+  userId: z.number().int().positive(),
+  bio: z
+    .string()
+    .min(10, "Bio must be at least 10 characters")
+    .max(1000)
+    .optional(),
+  resumeUrl: z.url("Invalid resume URL").optional(),
+  linkedinUrl: z.url("Invalid LinkedIn URL").optional(),
+  portfolioUrl: z.url("Invalid portfolio URL").optional(),
+});
+
+export const selectUserSchema = createSelectSchema(user);
+export const selectUserProfileSchema = createSelectSchema(userProfile);
+
+export const updateUserSchema = insertUserSchema.partial().omit({
+  id: true,
+  createdAt: true,
+  passwordHash: true,
+  updatedAt: true,
+  deletedAt: true,
+});
+
+export const updateUserProfileSchema = insertUserProfileSchema
+  .omit({ userId: true })
+  .extend({
+    educations: insertEducationsSchema
+      .omit({ userProfileId: true })
+      .array()
+      .default([]),
+    workExperiences: insertWorkExperiencesSchema
+      .omit({ userProfileId: true })
+      .array()
+      .default([]),
+    certifications: insertCertificationsSchema.array().default([]),
+  });
+
+// Type exports
+export type User = z.infer<typeof selectUserSchema>;
+export type NewUser = z.infer<typeof insertUserSchema>;
+export type UpdateUser = z.infer<typeof updateUserSchema>;
+export type UserProfile = z.infer<typeof selectUserProfileSchema>;
+export type NewUserProfile = z.infer<typeof insertUserProfileSchema>;
+export type UpdateUserProfile = z.infer<typeof updateUserProfileSchema>;
+
+// Get type with relations
+export type UserWithProfile = User & {
+  profile:
+    | (UserProfile & {
+        certifications: { certification: Certification }[] | null;
+        education: Education[] | null;
+        workExperiences: WorkExperience[] | null;
+      })
+    | null;
+};
 
 const idParamsSchema = z.object({
   id: z.string().regex(/^\d+$/, "Invalid work experience ID format"),
